Refactor Analogue Scamp5 things into its own package
